add module for player
